[MMCA-4810] Incorporate review comments, move GuidanceRow model to common models
