+++ app/viewmodels/PostponedVatViewModel.scala
@@ -26,6 +26,7 @@ import utils.Utils._
 import views.helpers.Formatters
 import views.html.components._
 import views.html.postponed_vat.{collapsible_statement_group, current_statement_row, download_link_pvat_statement}
+import common.GuidanceRow
 
 case class PvEmail(emailAddress: String, emailAddressHref: String)
