+++ app/viewmodels/common/GuidanceRow.scala
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package viewmodels
+package viewmodels.common
 
 import play.twirl.api.HtmlFormat
 
-case class GuidanceRow(h2Heading: HtmlFormat.Appendable, link: Option[HtmlFormat.Appendable] = None)
-
-case class GuidanceRowWithParagraph(
+case class GuidanceRow(
   h2Heading: HtmlFormat.Appendable,
   link: Option[HtmlFormat.Appendable] = None,
   paragraph: Option[HtmlFormat.Appendable] = None
